use shorter sd_send_bytes() routines

+++ BaS_gcc/include/sd_card.h
@@ -27,7 +27,7 @@ const uint32_t DSPI_DMCR_CONF = MCF_DSPI_DMCR_MSTR |	/* FireBee is DSPI master*/
 extern int sd_card_init(void);
 extern uint32_t sd_com(uint32_t data);
 extern void sd_card_idle(void);
-extern uint32_t sd_card_get_status(void);
+extern uint8_t sd_card_get_status(void);
 extern uint8_t sd_send_byte(uint8_t byte);
 extern uint16_t sd_send_word(uint16_t word);
